* 서버 코어 작업자 분리 및 진단 코드 정리
* 재전송 및 세션 해제 스레드 함수를 작업자 구현 파일로 이동

* MemoryTracer 보고서 생성을 스트림 기반으로 통합

* C4819 발생 소스를 UTF-8 BOM으로 정규화
