+++ MultiSocketRUDP/CoreTest/RUDPSessionManagerTest.cpp
@@ -1,4 +1,4 @@
-#include "PreCompile.h"
+﻿#include "PreCompile.h"
 #include <gtest/gtest.h>
 
 #include "MultiSocketRUDPCore.h"
