+++ MultiSocketRUDP/MultiSocketRUDPClient/RUDPClientCore.h
@@ -1,4 +1,4 @@
-#pragma once
+﻿#pragma once
 #include "NetServerSerializeBuffer.h"
 #include "BuildConfig.h"
 #include "../Common/etc/CoreType.h"
