+++ MultiSocketRUDP/MultiSocketRUDPServer/RIOManager.h
@@ -1,4 +1,4 @@
-#pragma once
+﻿#pragma once
 #include "IRIOManager.h"
 #include <MSWSock.h>
 #include <vector>
