+++ MultiSocketRUDP/MultiSocketRUDPServer/RUDPSessionManager.h
@@ -1,4 +1,4 @@
-#pragma once
+﻿#pragma once
 #include "RUDPSession.h"
 
 class ISessionDelegate;
