+++ MultiSocketRUDP/MultiSocketRUDPServer/SessionStateMachine.h
@@ -1,4 +1,4 @@
-#pragma once
+﻿#pragma once
 #include <atomic>
 #include "../Common/etc/CoreType.h"
